chore(symbol-usage): move `n usages` text to the end of the line instead of above
This way, we don't lose lines every time a function has some usages,
which was becoming a bit annoying.

I tried putting it on the sign column, and I kind of like it with just
the numvber, but then I have to activate 2 signcolumns if I want my
other git signs to show up in the same line, and it just becomes a bit
too wide I think.

So I'm leaving it on the right side for now, maybe it's not that much
visible but should be good enough.