feat(nvim): update neogen command
The reason is that it does not always pick up that you
are on top of a function, which is the 99% of the times
I will be using it.

So appending func makes it work always for this case.

In case I want to use it for other types, I can always
use the command manually.
Because that will hardly every happen, it's fine to just
have 1 keymap for this.